Revert "Any changes after remarks."
This reverts commit 927837d4
