+++ src/test/java/it_academy/context/StudentTest.java
@@ -1,7 +1,6 @@
 package it_academy.context;
 
 import it_academy.ConstantContainer;
-import it_academy.Types;
 import it_academy.strategy.cocreteStrtegies.FirstTypeStrategy;
 import it_academy.strategy.cocreteStrtegies.SecondTypeStrategy;
 import it_academy.strategy.cocreteStrtegies.ThirdTypeStrategy;
@@ -22,7 +21,7 @@ class StudentTest {
 
 	@BeforeAll
 	static void createInstance() {
-		student = new Student(Types.values()[TEST_TYPE_1], TEST_TALENT);
+		student = new Student(ConstantContainer.Types.values()[TEST_TYPE_1], TEST_TALENT);
 	}
 
 	@Test
@@ -36,13 +35,13 @@ class StudentTest {
 		Field field = clazz.getDeclaredField("calc");
 		field.setAccessible(true);
 
-		student.setType(Types.values()[TEST_TYPE_1]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_1]);
 		assertTrue(field.get(student) instanceof FirstTypeStrategy);
 
-		student.setType(Types.values()[TEST_TYPE_2]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_2]);
 		assertTrue(field.get(student) instanceof SecondTypeStrategy);
 
-		student.setType(Types.values()[TEST_TYPE_3]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_3]);
 		assertTrue(field.get(student) instanceof ThirdTypeStrategy);
 
 		field.setAccessible(false);
@@ -50,17 +49,17 @@ class StudentTest {
 
 	@Test
 	void testToString() {
-		student.setType(Types.values()[TEST_TYPE_1]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_1]);
 		assertEquals("Student{ talent=0.5; totalTime=396; type=1; typesOfTraining: [potok, razbor, practic]; time To Each Type Of Training=132}", student.toString());
-		student.setType(Types.values()[TEST_TYPE_2]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_2]);
 		assertEquals("Student{ talent=0.5; totalTime=792; type=2; typesOfTraining: [razbor, practic]; time To Each Type Of Training=396}", student.toString());
-		student.setType(Types.values()[TEST_TYPE_3]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_3]);
 		assertEquals("Student{ talent=0.5; totalTime=1188; type=3; typesOfTraining: [practic]; time To Each Type Of Training=1188}", student.toString());
 	}
 
 	@Test
 	void getTotalTime() {
-		student.setType(Types.values()[TEST_TYPE_1]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_1]);
 
 		int expected = (int) (ConstantContainer.TOTAL_LENGTH_OF_EDUCATION_PROCESS_FOR_TYPE_1_AND_TALENT_1
 				* ConstantContainer.X_LENGTH_OF_EDUCATIONAL_PROCESS_FOR_TYPE_1 / TEST_TALENT);
@@ -70,7 +69,7 @@ class StudentTest {
 
 	@Test
 	void getTimeToEachKindOfProcess() {
-		student.setType(Types.values()[TEST_TYPE_1]);
+		student.setType(ConstantContainer.Types.values()[TEST_TYPE_1]);
 
 		int expected = (int) (ConstantContainer.TOTAL_LENGTH_OF_EDUCATION_PROCESS_FOR_TYPE_1_AND_TALENT_1
 				* ConstantContainer.X_LENGTH_OF_EDUCATIONAL_PROCESS_FOR_TYPE_1
